Ensure maxAgeFn is called even on 5xx responses.
This commit fixes a problem where a 5xx response from the remote host
will cause an 'undefined' value to be cached without calling the
maxAgeFn first. This will cause clients of exp-fetch which provides a
maxAgeFn but doesn't set a maxAge property on the lru-cache to cache an
'undefined' value indefinitely.

The expectation is that maxAgeFn should be called even for 5xx responses
and this patch ensures that it is.