Fix bug when `cacheNotFound` was set to `true`
The old implementation handled `true` as a number, which resulted in it
being cast to 1 millisecond `maxAge`/TTL. Now the value of `true` uses
the "cache-control" header for cache age.

`cacheNotFound` is only tested for falsiness, so there is no point in
forcing it to be a boolean `false`.

Test that the cache has been hit (by checking the cache hit ratio).

Actually call `done` when the `expects` in the `then` functions throw
errors.