Remove `length` from `initLRUCache` settings
It is redundant and might actually have unexpected behavior if the item
that is put in the cache has a `length` attribute. Objects lacking this
attribute count as length 1, and the rest as the `length` attribute.

The `max` size in `lru-cache` 6.0.0 can be used as an item count
(length 1) or as a number of characters (or other arbitrary length
calculation). The previous implementation was a mix of both, if the
returned item happened to have a `length` attribute.

The intent here is to always count one item in the cache as 1 towards
the max. Since the default length in the NPM module is 1, we can just
skip providing it.

Also use the built-in `prune` function, instead of doing it with a dummy
forEach.

Move the most interesting code to the top.